Styled SignUpPage and SignUpForm components. Added missing fields.
