+++ src/components/user/SignUpForm/sign-up-form.scss
@@ -1,0 +1,12 @@
+.sign-up-form-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.sign-up-form {
+  div {
+    margin-top: 10%;
+  }
+}
