+++ src/locales/es.js
@@ -11,7 +11,9 @@ export default {
 
   'signup.title': 'REGISTRARSE',
   'signup.signin': 'Iniciar Sesión',
-  'signup.form.passconfirmation': 'Confirmación de Contraseña',
+  'signup.form.gender': 'Género',
+  'signup.form.username': 'Nombre de usuario',
+  'signup.form.passconfirmation': 'Repetir contraseña',
   'signup.form.confirmation.presence': 'Debe confirmar su contraseña para continuar',
   'signup.form.confirmation.equality':
     'Su confirmación de contraseña debe ser igual a la contraseña',
