+++ src/pages/SignUpPage/sign-up-page.scss
@@ -1,0 +1,8 @@
+.sign-up-page-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  width: 50%;
+  justify-content: center;
+  align-items: center;
+}
